+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:to_do_app/src/constants/theme/theme.dart';
-import './presentation/screens/home/home.dart';
+import 'presentation/screens/home/home_screen.dart';
 
 void main() {
   runApp(const ToDoApp());
@@ -9,7 +9,6 @@ void main() {
 class ToDoApp extends StatelessWidget {
   const ToDoApp({super.key});
 
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
