+++ lib/presentation/widgets/screen/screen.dart
@@ -8,22 +8,27 @@ class Screen extends StatelessWidget {
   const Screen({Key? key, required this.child, this.onPressedAction})
       : super(key: key);
 
+  FloatingActionButton? renderFloatinButton() => onPressedAction != null
+      ? FloatingActionButton(
+          onPressed: onPressedAction,
+          backgroundColor: Color(0xFF1294f2),
+          child: const Icon(Icons.add, color: Colors.white),
+        )
+      : null;
+
   @override
   Widget build(BuildContext context) {
     return AnnotatedRegion<SystemUiOverlayStyle>(
         value: SystemUiOverlayStyle.light,
-        child: Scaffold(
-            body: SafeArea(
-                child: Padding(
-              padding: EdgeInsets.fromLTRB(10, 20, 10, 15),
-              child: child,
-            )),
-            floatingActionButton: onPressedAction != null
-                ? FloatingActionButton(
-                    onPressed: onPressedAction,
-                    backgroundColor: Color(0xFF1294f2),
-                    child: const Icon(Icons.add, color: Colors.white),
-                  )
-                : null));
+        child: WillPopScope(
+          child: Scaffold(
+              body: SafeArea(
+                  child: Padding(
+                padding: EdgeInsets.fromLTRB(10, 20, 10, 15),
+                child: child,
+              )),
+              floatingActionButton: renderFloatinButton()),
+          onWillPop: () async => false,
+        ));
   }
 }
