+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   intl: ^0.18.0
+  table_calendar: ^3.0.9
 
 dev_dependencies:
   flutter_test:
